MAIN update mockserver to 5.11.1 with a new http wait strategy
fixes #2984, #2710

in https://github.com/mock-server/mockserver/commit/9f27af00143a2418e0afe3d060f9e9efe233af3a mockserver introduced a GET http healthcheck (what was released in 5.9.0) and since 5.11.0 uses distroless image, with the default wait strategy, we hit into #3317

As the 5.11.0+ is actually a smaller image, more secure and 5.9.0 was released almost a year ago (which means users with 5.9.0+ versions would be able to update testcontainers without changes), I think worth bumping the base version.

For the users on old versions the old behaviour could be set with the default waiting strategy back
`.waitingFor(Wait.defaultWaitStrategy())`.

This PR also brings two small changes to be complete - mockserver official image now is `mockserver/mockserver` (but I added previous as a compatible).
And added small note to the docs regarding matched version of a client, as anyway had to change docs after test adjustments (wasn't able to test properly with rule in the same class as old version).
