+++ docs/modules/mockserver.md
@@ -7,13 +7,13 @@ Mock Server can be used to mock HTTP services by matching requests against user-
 The following example shows how to start Mockserver.
 
 <!--codeinclude-->
-[Creating a MockServer container](../../modules/mockserver/src/test/java/org/testcontainers/containers/MockServerContainerTest.java) inside_block:creatingProxy
+[Creating a MockServer container](../../modules/mockserver/src/test/java/org/testcontainers/containers/MockServerContainerRuleTest.java) inside_block:creatingProxy
 <!--/codeinclude-->
 
 And how to set a simple expectation using the Java MockServerClient.
 
 <!--codeinclude-->
-[Setting a simple expectation](../../modules/mockserver/src/test/java/org/testcontainers/containers/MockServerContainerTest.java) inside_block:testSimpleExpectation
+[Setting a simple expectation](../../modules/mockserver/src/test/java/org/testcontainers/containers/MockServerContainerRuleTest.java) inside_block:testSimpleExpectation
 <!--/codeinclude-->
 
 ## Adding this module to your project dependencies
@@ -33,3 +33,5 @@ testCompile "org.testcontainers:mockserver:{{latest_version}}"
 </dependency>
 ```
 
+Additionally, don't forget to add a [client dependency `org.mock-server:mockserver-client-java`](https://search.maven.org/search?q=mockserver-client-java) 
+to be able to set expectations, it's not provided by the testcontainers module. Client version should match to the version in a container tag.
