+++ modules/mockserver/src/main/java/org/testcontainers/containers/MockServerContainer.java
@@ -1,13 +1,14 @@
 package org.testcontainers.containers;
 
 import lombok.extern.slf4j.Slf4j;
+import org.testcontainers.containers.wait.strategy.Wait;
 import org.testcontainers.utility.DockerImageName;
 
 @Slf4j
 public class MockServerContainer extends GenericContainer<MockServerContainer> {
 
-    private static final DockerImageName DEFAULT_IMAGE_NAME = DockerImageName.parse("jamesdbloom/mockserver");
-    private static final String DEFAULT_TAG = "mockserver-5.5.4";
+    private static final DockerImageName DEFAULT_IMAGE_NAME = DockerImageName.parse("mockserver/mockserver");
+    private static final String DEFAULT_TAG = "mockserver-5.11.1";
 
     @Deprecated
     public static final String VERSION = DEFAULT_TAG;
@@ -33,7 +34,10 @@ public class MockServerContainer extends GenericContainer<MockServerContainer> {
     public MockServerContainer(DockerImageName dockerImageName) {
         super(dockerImageName);
 
-        dockerImageName.assertCompatibleWith(DEFAULT_IMAGE_NAME);
+        dockerImageName.assertCompatibleWith(DEFAULT_IMAGE_NAME, DockerImageName.parse("jamesdbloom/mockserver"));
+
+        withEnv("MOCKSERVER_LIVENESS_HTTP_GET_PATH", "/mockserver/health");
+        waitingFor(Wait.forHttp("/mockserver/health").forStatusCode(200));
 
         withCommand("-logLevel INFO -serverPort " + PORT);
         addExposedPorts(PORT);
