+++ modules/mockserver/src/test/java/org/testcontainers/containers/MockServerContainerTest.java
@@ -1,56 +1,21 @@
 package org.testcontainers.containers;
 
-import lombok.Cleanup;
-import org.junit.Rule;
 import org.junit.Test;
 import org.mockserver.client.MockServerClient;
+import org.testcontainers.containers.wait.strategy.Wait;
 import org.testcontainers.utility.DockerImageName;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.net.URL;
-import java.net.URLConnection;
+import java.io.FileNotFoundException;
 
-import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.mockserver.model.HttpRequest.request;
 import static org.mockserver.model.HttpResponse.response;
 import static org.rnorth.visibleassertions.VisibleAssertions.assertThat;
+import static org.rnorth.visibleassertions.VisibleAssertions.assertThrows;
 
 public class MockServerContainerTest {
 
-    public static final DockerImageName MOCKSERVER_IMAGE = DockerImageName.parse("jamesdbloom/mockserver:mockserver-5.5.4");
-
-    // creatingProxy {
-    @Rule
-    public MockServerContainer mockServer = new MockServerContainer(MOCKSERVER_IMAGE);
-    // }
-
-    private static String responseFromMockserver(MockServerContainer mockServer, String path) throws IOException {
-        URLConnection urlConnection = new URL(mockServer.getEndpoint() + path).openConnection();
-        @Cleanup BufferedReader reader = new BufferedReader(new InputStreamReader(urlConnection.getInputStream()));
-        return reader.readLine();
-    }
-
-    @Test
-    public void shouldReturnExpectation() throws Exception {
-        // testSimpleExpectation {
-        new MockServerClient(mockServer.getHost(), mockServer.getServerPort())
-            .when(request()
-                .withPath("/person")
-                .withQueryStringParameter("name", "peter"))
-            .respond(response()
-                .withBody("Peter the person!"));
-
-        // ...a GET request to '/person?name=peter' returns "Peter the person!"
-        // }
-
-        assertThat("Expectation returns expected response body",
-            responseFromMockserver(mockServer, "/person?name=peter"),
-            containsString("Peter the person")
-        );
-    }
+    public static final DockerImageName MOCKSERVER_IMAGE = DockerImageName.parse("mockserver/mockserver:mockserver-5.11.1");
 
     @Test
     public void shouldCallActualMockserverVersion() throws Exception {
@@ -65,8 +30,20 @@ public class MockServerContainerTest {
                 .respond(response().withBody(expectedBody));
 
             assertThat("MockServer returns correct result",
-                responseFromMockserver(mockServer, "/hello"),
+                SimpleHttpClient.responseFromMockserver(mockServer, "/hello"),
                 equalTo(expectedBody)
+            );
+        }
+    }
+
+    @Test
+    public void oldVersionRequiresDefaultWaitStrategy() throws Exception {
+        DockerImageName dockerImageName = DockerImageName.parse("jamesdbloom/mockserver").withTag("mockserver-5.5.4");
+        try (MockServerContainer mockServer = new MockServerContainer(dockerImageName).waitingFor(Wait.defaultWaitStrategy())) {
+            mockServer.start();
+
+            assertThrows("expected not found response (as we can't use old client to set expectations properly)", FileNotFoundException.class,
+                () -> SimpleHttpClient.responseFromMockserver(mockServer, "/hello")
             );
         }
     }
